Fix button issues and move horizontal scroll fix
Introduces a btn_canvas label as the parent for the buttons, which both fixes the button's undrawn background looking bad and fixes the wobble when resizing from the top border. This fix removes the need for layered window attributes, so that has been removed, improving performance. Also moved the fix for the vertical line/horizontal scrollbar issue in the listview to the recently implemented timer, eliminating yet another paint counter.