Make process list icons DPI aware and tweak fonts
The icons in the process list now scale when the DPI changes. Making it change size immediately when changing DPI required a lot of work. The Icons have to be cached in a shared variable now to allow for scaling the icons without refreshing the window list and without losing quality.
Also, changed the font quality from DRAFT to DEFAULT.