Change image list handling method
Adds a shared variable that stores the image list and uses that instead of the returned handle to properly drop the old image list when refreshing. This also ensures that the image list is correctly dropped when the app is closed, which I was not certain occurred before.
Also, add condition to the refresh without scanning code so that it does not populate the new image list if icon use is disabled.